+++ spec/tmdb_spec.rb
@@ -54,5 +54,35 @@ describe TMDb do
         a_request(:get, expected_url).with(:query => expected_query).should have_been_made
       end
     end
+    context 'when not configured with a cache' do
+      it 'makes an HTTP request once per call with same parameters' do
+        3.times { TMDb.get_api_response(example_path, example_params) }
+        a_request(:get, expected_url).with(:query => expected_query).should have_been_made.times(3)
+      end
+    end
+    context 'when configured with a cache' do
+      it 'makes a single HTTP request per call with same parameters' do
+        [1, 2].each do |n|
+          stub_request(:get, "http://api.themoviedb.org/3/path#{n}").
+            with(:query => hash_including({})).
+            to_return(:body => example_response)
+        end
+        TMDb.configure {|config| config.cache = ActiveSupport::Cache::MemoryStore.new }
+        3.times do
+          TMDb.get_api_response('path1')
+          TMDb.get_api_response('path2')
+          TMDb.get_api_response('path1', 'param' => '1')
+          TMDb.get_api_response('path1', 'param' => '2')
+        end
+        a_request(:get, 'http://api.themoviedb.org/3/path1')
+          .with(:query => { :api_key => example_api_key }).should have_been_made.times(1)
+        a_request(:get, 'http://api.themoviedb.org/3/path2')
+          .with(:query => { :api_key => example_api_key }).should have_been_made.times(1)
+        a_request(:get, 'http://api.themoviedb.org/3/path1').
+          with(:query => hash_including({ 'param' => '1' })).should have_been_made.times(1)
+        a_request(:get, 'http://api.themoviedb.org/3/path1').
+          with(:query => hash_including({ 'param' => '2' })).should have_been_made.times(1)
+      end
+    end
   end
 end